First iteration
Current edit is functional running with "yarn test" command, and some elements of the web page are being tested properly. I'm running into an error with one of my tests, but I think it must be a misunderstanding of some of the driver.io syntax. I'll work on it later.